Change IEnumerable to IIterable and modify Array<T, Size>

+++ Source/Number.cpp
@@ -6,17 +6,17 @@ namespace Stella
 	                                 Fields
 	-----------------------------------------------------------------------*/
 
-	const Int8 Number::Int8MaxValue = 0x7F;
-	const Int8 Number::Int8MinValue = 0x80;
+	const Int8 Number::Int8MaxValue   = (Int8) 0x7F;
+	const Int8 Number::Int8MinValue   = (Int8) 0x80;
 
-	const Int16 Number::Int16MaxValue = 0x7FFF;
-	const Int16 Number::Int16MinValue = 0x8000;
+	const Int16 Number::Int16MaxValue = (Int16) 0x7FFF;
+	const Int16 Number::Int16MinValue = (Int16) 0x8000;
 
-	const Int32 Number::Int32MaxValue = 0x7FFFFFFF;
-	const Int32 Number::Int32MinValue = 0x80000000;
+	const Int32 Number::Int32MaxValue = (Int32) 0x7FFFFFFF;
+	const Int32 Number::Int32MinValue = (Int32) 0x80000000;
 
-	const Int64 Number::Int64MaxValue = 0x7FFFFFFFFFFFFFFF;
-	const Int64 Number::Int64MinValue = 0x8000000000000000;
+	const Int64 Number::Int64MaxValue = (Int64) 0x7FFFFFFFFFFFFFFF;
+	const Int64 Number::Int64MinValue = (Int64) 0x8000000000000000;
 
 	/*-----------------------------------------------------------------------
 		                                 Functions
